requirements.txt
librosa==0.9.2
matplotlib==3.6.1
matplotx[all]~=0.3.10
Pillow~=9.2.0
numpy~=1.23.4